moves slack to common runtime, disables sending slack messages automatically
